feat : 추상 클래스 생성 및 상속 구현
