+++ README.md
@@ -14,6 +14,7 @@
 - [ ] 모든 원시값과 문자열을 포장한다.
 - [ ] 일급 컬렉션을 쓴다.
 - [ ] 객체지향생활체조를 모두 만족한다. (9번은 쓰지 않는다 말고 지양한다 수준으로 맞춰본다.)
+- [ ] 모든 클래스에 abstract 혹은 final을 추가한다.
 
 ## ****추가된 요구 사항****
 
@@ -27,20 +28,21 @@
     - [ ] 21점을 넘지 않으면 계속해서 카드를 뽑을 수 있다.
     - [ ] 딜러와 플레이어 중 카드의 합이 21 또는 21에 가장 가까운 숫자를 가지는 쪽을 판별한다.
     - [ ] 게임을 시작하면 플레이어와 딜러에게 두 장의 카드를 지급한다.
-- [ ] 플레이어 : 인터페이스
+
+- [ ] 플레이어 : 추상클래스
     - [x] 자신의 카드 리스트를 보여준다.
-- [ ] 딜러
-    - [ ] 플레이어랑 같은데, 카드를 한장만 공개함
-    - [x] 자신의 카드 리스트를 보여준다.
-    - [ ] 16 이하면 무조건 카드를 한장 더 뽑음 / 17점 이상이면 추가 X
-- [ ] 참가자
     - [x] 이름
     - [x] 카드 일급컬렉션
         - [x] 카드의 합을 도출
         - [x] Ace는 1 또는 11로 계산한다.
-            - [x] Ace를 제외한 카드들의 합이 10 이하면 Ace를 11로 계산한다.
-    - [ ] 21미만이면 카드를 더 뽑을 건지 묻는다. (21이상이면 묻지않는다.)
-    - [x] 자신의 카드 리스트를 보여준다.
+    - [x] Ace를 제외한 카드들의 합이 10 이하면 Ace를 11로 계산한다.
+
+- [ ] 딜러
+    - [ ] 플레이어랑 같은데, 카드를 한장만 공개함
+    - [ ] 16 이하면 무조건 카드를 한장 더 뽑음 / 17점 이상이면 추가 X
+
+- [ ] 참가자
+  - [ ] 21미만이면 카드를 더 뽑을 건지 묻는다. (21이상이면 묻지않는다.)
 - [x]  덱
     - [x] RandomGenerator에게 받은 셔플된 카드들을 deque에 저장
     - [x] 덱에서 카드를 뽑아서 리턴한다.
