+++ src/main/java/domain/Player.java
@@ -2,7 +2,24 @@ package domain;
 
 import java.util.List;
 
-@FunctionalInterface
-public interface Player {
-    List<Card> displayCards();
+public abstract class Player {
+    private final Name name;
+    private final Cards cards;
+
+    Player(final Name name, final Cards cards) {
+        this.name = name;
+        this.cards = cards;
+    }
+
+    public List<Card> displayCards() {
+        return cards.displayCards();
+    }
+
+    public void takeCard(final Card card) {
+        cards.takeCard(card);
+    }
+
+    public int getScore() {
+        return cards.getScore();
+    }
 }
